moved measurements to their own file
